feat: update sidebar component with dynamic routing and add Navbar to layout
