+++ src/app/[year]/layout.tsx
@@ -1,5 +1,6 @@
 import BackgroundLayout from "./(components)/common/BackgroundLayout";
 import "../../styles/globals.css";
+import Navbar from "./(components)/common/Navbar";
 export default function YearSeasonLayout({
   children,
 }: {
@@ -7,7 +8,10 @@ export default function YearSeasonLayout({
 }) {
   return (
     <>
-      <BackgroundLayout>{children}</BackgroundLayout>
+      <BackgroundLayout>
+        <Navbar />
+        {children}
+      </BackgroundLayout>
     </>
   );
 }
